-декомпозирует компоненты на странице favorites-page
